src/processors/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
python-snappy==0.6.1
